requirements.txt
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Faker==22.6.0
fastapi==0.109.0
Flask==3.0.0
h11==0.14.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
numpy==1.26.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.4
pluggy==1.3.0
psycopg2-binary==2.9.9
pydantic==2.6.0
pydantic_core==2.16.1
pyodbc==5.0.1
PySocks==1.7.1
pytest==7.4.4
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.35.1
trio==0.23.2
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.27.0.post1
Werkzeug==3.0.1
wsproto==1.2.0